AV-54034 top nav highlighting WIP
This is more complicated than it looks, partly because the Nav data-structure
is a bit odd (e.g. the Home icon isn't generated from the Home configuration for...
reasons?)

This involves grubbing around in the site keys, and is therefore dependent
on magic globals that are created while running templating code...

(e.g. it works, but will require some documentation or rework, as will be fun
to debug in future)